backend/requirements.txt
# Screenplay parsing
pdfplumber
beautifulsoup4
chardet
screenpy @ git+https://github.com/drwiner/ScreenPy.git

# TMDB metadata
requests
python-dotenv

# Retrieval
sentence-transformers
faiss-cpu
rank-bm25

# API
fastapi
uvicorn[standard]
